+++ recipes/default.rb
@@ -54,7 +54,7 @@ if fqdn
 
     execute "hostname #{fqdn}" do
       only_if { node['fqdn'] != fqdn }
-      notifies :reload, 'ohai[reload]', :immediately
+      notifies :reload, 'ohai[reload_hostname]', :immediately
     end
 
   when 'centos', 'redhat', 'amazon', 'scientific'
@@ -65,7 +65,7 @@ if fqdn
         file.search_file_replace_line('^HOSTNAME', "HOSTNAME=#{fqdn}")
         file.write_file
       end
-      notifies :reload, 'ohai[reload]', :immediately
+      notifies :reload, 'ohai[reload_hostname]', :immediately
     end
     # this is to persist the correct hostname after machine reboot
     sysctl = '/etc/sysctl.conf'
@@ -76,28 +76,30 @@ if fqdn
                                      "kernel.hostname=#{hostname}")
         file.write_file
       end
-      notifies :reload, 'ohai[reload]', :immediately
+      notifies :reload, 'ohai[reload_hostname]', :immediately
     end
     execute "hostname #{hostname}" do
       only_if { has_new_hostname }
-      notifies :reload, 'ohai[reload]', :immediately
+
+      notifies :reload, 'ohai[reload_hostname]', :immediately
     end
-    if has_new_hostname
-      service 'network' do
-        action :restart
-      end
+    service 'network' do
+      action :restart
+      only_if { has_new_hostname }
+      notifies :reload, 'ohai[reload_hostname]', :immediately
     end
 
   else
     file '/etc/hostname' do
       content "#{hostname}\n"
       mode '0644'
-      notifies :reload, 'ohai[reload]', :immediately
+      notifies :reload, 'ohai[reload_hostname]', :immediately
     end
 
     execute "hostname #{hostname}" do
       only_if { has_new_hostname }
-      notifies :reload, 'ohai[reload]', :immediately
+
+      notifies :reload, 'ohai[reload_hostname]', :immediately
     end
   end
 
@@ -112,10 +114,11 @@ if fqdn
     hostname fqdn
     aliases [hostname]
     action :create
-    notifies :reload, 'ohai[reload]', :immediately
+    notifies :reload, 'ohai[reload_hostname]', :immediately
   end
 
-  ohai 'reload' do
+  ohai 'reload_hostname' do
+    plugin 'hostname'
     action :nothing
   end
 else