+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 # Change History
 
+
+Unreleased
+==========
+ - #27: Update default recipe to reload only the hostname plugin instead of ohai (Jonathan Serafini)
+
 0.3.0
-=====
  - Fixed (and tested) FreeBSD support
  - #17: added support for RedHat & CentOS (Damien Roche, Marta Paciorkowska)
  - added instructions on manual testing with reboot (Marta Paciorkowska)